+++ shopping-client/src/main/java/br/com/rocha/shoppingclient/dto/ProductDTO.java
@@ -1,9 +1,7 @@
-package br.com.rocha.productapi.dto;
+package br.com.rocha.shoppingclient.dto;
 
-import br.com.rocha.productapi.model.Category;
-import br.com.rocha.productapi.model.Product;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 
 public class ProductDTO {
 
@@ -21,46 +19,40 @@ public class ProductDTO {
     public String getProductIdentifier() {
         return productIdentifier;
     }
+
     public void setProductIdentifier(String productIdentifier) {
         this.productIdentifier = productIdentifier;
     }
+
     public String getName() {
         return name;
     }
+
     public void setName(String name) {
         this.name = name;
     }
+
     public String getDescription() {
         return description;
     }
+
     public void setDescription(String description) {
         this.description = description;
     }
+
     public Float getPrice() {
         return price;
     }
+
     public void setPrice(Float price) {
         this.price = price;
     }
+
     public CategoryDTO getCategoryDTO() {
         return category;
     }
+
     public void setCategoryDTO(CategoryDTO category) {
         this.category = category;
     }
-
-    public static ProductDTO convert(Product product){
-
-        ProductDTO productDTO = new ProductDTO();
-        productDTO.setName(product.getName());
-        productDTO.setPrice(product.getPrice());
-        productDTO.setProductIdentifier(product.getProductIdentifier());
-        productDTO.setDescription(product.getDescription());
-
-        if(product.getCategory() != null){
-            productDTO.setCategoryDTO(CategoryDTO.convert(product.getCategory()));
-        }
-        return productDTO;
-    }
-    
 }
